--- src/services/shared/http-response.ts
@@ -1,1 +1,0 @@
-export { readJsonResponse } from '../../shared/http-response';
